requirements.txt
Mezzanine==4.1.0
MySQL-python==1.2.5
django-dotenv==1.4.1
